pyproject.toml
[project]
name = "wettest"
requires-python = "==3.12"
version = "0.0.0"
dependencies = [
    "pyyaml==6.0.1",
    "openai==1.56.2",
    "types-PyYAML==6.0.12.11",
    "mypy==1.4.1",
    "requests>=2.32.3,<3",
]

[tool.pixi.project]
name = "wettest"
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
wettest = { path = ".", editable = true }

[tool.pixi.tasks]
bot = "OPENAI_API_KEY=$(cat openai.key) ./src/bot.py"
mypy = "mypy ."
